backend/metrics/requirements.txt
supabase-py
pandas
pyActigraphy
neurokit2
pyhrv
